Separate operator generation and building logic in scripts
The scripts are: codegen-operator and build-operator, we rely
on a stable dapper based environment to be able to generate
and build across the team without issues.

The dapper image now installs operator-sdk from an specific
master commit it's know for us to work.

To regenerate the operators/go/submariner-operator:

    make codegen-operator

To build the operator container image:

    make build-operator