Add waits for secret creation before verifications
Per feedback from Noam, and I've also seen it myself, the timing can be
such that the secret verifications fail because the secret doesn't quite
exist yet.

There is no "condition:Ready" on Secret objects as there are is on Pods,
so we can't use the normal `wait --for=condition=Ready secrets --all`
here.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>