Add kubectl to the route-agent image
In order to ensure that host path mounts don't interfere with socket
creation, the route agent pods need to wait for the node to be ready
(which includes CNI readiness) before starting. The simplest way to do
that is to ship kubectl in the container image; to ensure it is built
from scratch, this adds the minimal command wrapper in a separate
module, and builds it for use in the image container.

This will be put to use in the operator; the corresponding change will
have full details of the issue being fixed.

Signed-off-by: Stephen Kitt <skitt@redhat.com>